server/requirements.txt
# --- Core framework ---
Flask==3.0.3
Flask-Cors==4.0.1
Werkzeug==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe>=3.0.2   # ensures prebuilt wheels for 3.13

# --- Web servers ---
gunicorn==22.0.0     # production (Linux/macOS)
waitress==3.0.2      # optional (Windows/dev)

# --- Database ---
# Upgrade or switch driver for Python 3.13
psycopg2-binary>=2.9.10
# Alternatively:
# psycopg[binary]>=3.1.19

# --- Data models & utilities ---
pydantic==2.8.2
annotated-types==0.7.0
pydantic_core==2.20.1
packaging==25.0
python-dotenv==1.1.1

# --- CLI & signals ---
click==8.1.8
blinker==1.9.0
colorama==0.4.6

# --- Compatibility shims (can drop on 3.13) ---
# typing_extensions and importlib_metadata are built in now
# typing_extensions>=4.15.0
# importlib_metadata>=8.7.0

# --- Misc ---
zipp==3.23.0
